elle.Buffer: use size_t instead of uint64_t
Let's stick to the regular type for sizes.  A quick experiment with
ssize_t generated many warnings in our code, and in the interfaces
with other libs to which we hand over Buffer's content and size.

+++ src/elle/IOStream.hh
@@ -46,7 +46,7 @@ namespace elle
     : public std::streambuf
   {
   public:
-    using Size = uint64_t;
+    using Size = size_t;
     StreamBuffer();
 
     ~StreamBuffer() override;
